Relocated .pyi files to relevant packages
