+++ airflow/providers/microsoft/azure/operators/wasb_delete_blob.pyi
@@ -1,0 +1,40 @@
+#
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+#
+from typing import Optional, Any
+
+from airflow.models import BaseOperator
+
+
+class WasbDeleteBlobOperator(BaseOperator):
+    """
+    A stub file to suppress MyPy issues due to not supplying
+    mandatory parameters to the operator
+    """
+
+    def __init__(
+        self,
+        *,
+        container_name: Optional[str] = None,
+        blob_name: Optional[str] = None,
+        wasb_conn_id: str = 'wasb_default',
+        check_options: Any = None,
+        is_prefix: bool = False,
+        ignore_if_missing: bool = False,
+        **kwargs,
+    ) -> None: ...
